% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream_processing.R
\name{add_snv_group_to_sobj}
\alias{add_snv_group_to_sobj}
\title{Add SNV Group to Seurat Object}
\usage{
add_snv_group_to_sobj(
  sobject,
  snv_group_file,
  new_column,
  cell_group = "used_clusters"
)
}
\arguments{
\item{sobject}{A Seurat object to which the SNV group information will be added.}

\item{snv_group_file}{A file path to a tab-separated values (TSV) file
containing the SNV group information. The file should have two columns:
"cluster" and "dist_group". This is output from `get_snp_tree()`.}

\item{new_column}{A string specifying the name of the new column to be added
to the Seurat object's metadata.}

\item{cell_group}{A string specifying the name of the column in the Seurat
object's metadata that contains the cluster information. Default is
"used_clusters". This is the same column used in `get_snp_tree()` in the
table provided to the cellid_bam_table argument.}
}
\value{
The Seurat object with the new column added to its metadata.
}
\description{
This function adds a new column to the metadata of a Seurat object based on
a provided SNV (single nucleotide variant) group file. The new column will
contain the SNV group information for each cluster in the Seurat object.
}
\examples{
\dontrun{
sobj <- add_snv_group_to_sobj(
    sobject = seurat_object,
    snv_group_file = "path/to/snv_group_file_5_dist.txt",
    new_column = "snv_group_5"
)
}

}
