% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_variability.R
\name{calc_variant_variability}
\alias{calc_variant_variability}
\title{Use Merged BCF File to Quantify Variant Variability Across Groups}
\usage{
calc_variant_variability(
  merged_bcf_file,
  group_1,
  group_2,
  output_file = paste0(merged_bcf_file, "_variable.tsv"),
  min_snvs_per_cluster = 250,
  max_prop_missing_at_site = 0.5,
  verbose = TRUE,
  job_base = "batch",
  account = "gdrobertslab",
  log_base = "slurm_log_var",
  temp_dir = tempdir(),
  submit = TRUE,
  other_job_header_options = "",
  other_batch_options = "",
  sge_q = "all.q",
  sge_parallel_environment = "smp",
  job_scheduler = "slurm"
)
}
\arguments{
\item{merged_bcf_file}{Character. Path to the merged BCF file.}

\item{group_1}{Comma delimited string of multiple cluster names for
group 1. No spaces.}

\item{group_2}{Comma delimited string of multiple cluster names for
group 2. No spaces.}

\item{output_file}{Character. Path to the output file cross group . Default
is merged_bcf_file + "_variable.tsv".}

\item{min_snvs_per_cluster}{Numeric. Minimum number of SNVs per cluster.
Default is 250.}

\item{max_prop_missing_at_site}{The maximum proportion of missing data
allowed at a site.}

\item{verbose}{Whether to print out verbose output or not.}

\item{job_base}{The prefix to use with the sbatch job file.}

\item{account}{The hpc account to use.}

\item{log_base}{The directory to write job output files to.}

\item{temp_dir}{The directory to write temporary files to. This should be
unique to each sample run to avoid file collisions.}

\item{submit}{Whether to submit the sbatch jobs to the cluster or not.}

\item{other_job_header_options}{Other options to pass to the batch job
header. This should be a list of strings. Do not put "#SBATCH" or "#$" in
front of the strings.}

\item{other_batch_options}{Other options to pass to the batch job. These
optional strings will not be part of the header and will be placed verbatim
after the header. This is useful for passing module load commands, adding
to the path or other system-specific commands. Should be a list of strings.}

\item{sge_q}{Character string specifying the cluster queue when using sge as
the job scheduler.}

\item{sge_parallel_environment}{Character string specifying the parallel
environment (part of the -pe argument in the header) when using sge as the
job scheduler}

\item{job_scheduler}{The job scheduler to use. One of "slurm", "sge" or
"bash". If specifying "sge", sge_q and sge_parallel should be specified if
defaults are not correct.}
}
\value{
The result of the batch job submission.
}
\description{
This function quantifies the variability of variants across two groups of
clusters in a merged BCF file. It uses a Python script to calculate the
variant variability and returns the results in a specified output file.
}
\details{
The output file has the following columns:
- chr: Chromosome
- pos: Position of the variant
- alt_gt: Alternate genotype
- group_1_within_dist: Genetic distance within group 1
- group_2_within_dist: Genetic distance within group 2
- cross_group_dist: Genetic distance between groups
- cross_minus_within_dist: Difference between cross group distance and
 within group distances
}
\examples{
\dontrun{
  calc_variant_variability(
      merged_bcf_file = "path/to/merged.bcf",
      output_file = "variable_variants.tsv"
  )
}

}
