% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_variability.R
\name{call_variants_single_cell}
\alias{call_variants_single_cell}
\title{Call Variants on Each Cell From Select Variants}
\usage{
call_variants_single_cell(
  bam_to_use,
  variable_variants_file,
  cell_barcodes,
  output_file = "cell_variant_calls.tsv",
  ref_fasta,
  min_depth = 1,
  variant_diff_cutoff = 1,
  ploidy,
  job_base = "batch",
  account = "gdrobertslab",
  log_base = "slurm_log_var",
  temp_dir = tempfile(pattern = "tempdir"),
  submit = TRUE,
  other_job_header_options = "",
  other_batch_options = "",
  sge_q = "all.q",
  sge_parallel_environment = "smp",
  job_scheduler = "slurm"
)
}
\arguments{
\item{bam_to_use}{Character. Path to the BAM file to use for variant calling.}

\item{variable_variants_file}{Path to the location of the file produced by
calc_variant_variability().}

\item{cell_barcodes}{A single string with all cell barcodes to look for in
the bam file. Each barcode should be separated by a space.}

\item{output_file}{Character. Path to the output file of per-cell variant
calls.}

\item{ref_fasta}{The path to the reference fasta file.}

\item{min_depth}{The minimum depth to use when calling SNPs. Can be a vector
of numbers.}

\item{variant_diff_cutoff}{Numeric threshold to use to selct variants from
variable_variants_file. Any variant with cross_minus_within_dist greater
or equal to the cutoff will be used.}

\item{ploidy}{Either a file path to a ploidy file, or a string indicating
the ploidy. "GRCh37" is hg19, "GRCh38" is hg38, "mm10" is mm10.
"mm10_hg19" and "mm10_hg38" are our custom mixed species reference with
species prefixes on chromosomes. --ploidy-file option description here:
"https://samtools.github.io/bcftools/bcftools.html"}

\item{job_base}{The prefix to use with the sbatch job file.}

\item{account}{The hpc account to use.}

\item{log_base}{The directory to write job output files to.}

\item{temp_dir}{The directory to write temporary files to. This should be
unique to each sample run to avoid file collisions.}

\item{submit}{Whether to submit the sbatch jobs to the cluster or not.}

\item{other_job_header_options}{Other options to pass to the batch job
header. This should be a list of strings. Do not put "#SBATCH" or "#$" in
front of the strings.}

\item{other_batch_options}{Other options to pass to the batch job. These
optional strings will not be part of the header and will be placed verbatim
after the header. This is useful for passing module load commands, adding
to the path or other system-specific commands. Should be a list of strings.}

\item{sge_q}{Character string specifying the cluster queue when using sge as
the job scheduler.}

\item{sge_parallel_environment}{Character string specifying the parallel
environment (part of the -pe argument in the header) when using sge as the
job scheduler}

\item{job_scheduler}{The job scheduler to use. One of "slurm", "sge" or
"bash". If specifying "sge", sge_q and sge_parallel should be specified if
defaults are not correct.}
}
\value{
The result of the batch job submission. The data is printed to the
file specified by output_file.
}
\description{
This function calls variants on each cell from the selected variants.
}
\examples{
\dontrun{
  call_variants_single_cell(
      bam_to_use = "path/to/merged.bcf",
      variable_variants_file,
      cell_barcodes,
      ref_fasta,
      ploidy,
      output_file = "variable_variants.tsv"
  )
}

}
