% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream_processing.R
\name{label_tumor_cells}
\alias{label_tumor_cells}
\title{Label Tumor Groups in a Seurat Object Using SNV Groups}
\usage{
label_tumor_cells(
  sobject,
  cell_group = "used_clusters",
  snv_group_col,
  normal_clusters,
  tumor_call_column = "snv_tumor_call"
)
}
\arguments{
\item{sobject}{A Seurat object containing single-cell RNA-seq data.}

\item{cell_group}{A character string specifying the column name in the
Seurat object metadata that contains the cluster information. Default is
"used_clusters".}

\item{snv_group_col}{A character string specifying the column name in the
Seurat object metadata that contains the SNV group information derived from
`add_snv_group_to_sobj()`.}

\item{normal_clusters}{A vector of cluster identifiers that are considered
normal (non-tumor) clusters. These clusters should be present in the
`cell_group` column of the Seurat object metadata.}

\item{tumor_call_column}{A character string specifying the column name in the
Seurat object metadata where the tumor call labels will be stored. Default
is "snv_tumor_call".}
}
\value{
A Seurat object with an additional column in the metadata indicating
  whether each cell belongs to a normal, tumor, or unknown group. The
  "unknown" cells indicate there was no snv tumor call for those cells.
}
\description{
This function labels tumor cells in a Seurat object based on the specified
single nucleotide variant (SNV) group column and normal clusters.
}
\details{
If tumor_call_column already exists in the Seurat object metadata,
  it will be overwritten. The function identifies normal SNV groups
  based on the provided normal clusters and labels cells accordingly. If no
  normal clusters are found, the function will return an error message.
}
\examples{
\dontrun{
  sobject <- label_tumor_cells(
      sobject = seurat_obj,
      cell_group = "used_clusters",
      snv_group_col = "snv_group_20",
      normal_clusters = c("cluster1", "cluster2"),
      tumor_call_column = "snv_tumor_call"
  )
}

}
