% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_data.R
\name{prep_demo_data}
\alias{prep_demo_data}
\title{Prepare Demo Data}
\usage{
prep_demo_data(data_loc = find.package("scanBit"))
}
\arguments{
\item{data_loc}{Directory in which to download and extract the demo data.}
}
\description{
Downloads the demo data and updates the cell barcode table so its BAM file
paths are relative to the installed scanBit package directory. Data are
downloaded from https://doi.org/10.6084/m9.figshare.33399454.
}
