% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrrSnvs-package.R
\docType{package}
\name{rrrSnvs-package}
\alias{rrrSnvs}
\alias{rrrSnvs-package}
\title{rrrSnvs: A Set of Tools to Analyze SNV Data in Single Cell RNA-seq}
\description{
This is a set of functions to analyze variants in single cell data. The primary purpose is to help distinguish tumor cells. Primarily designed for use by folks within the Roberts lab but may be of use to others.
}
\keyword{internal}
