% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{use_job_template}
\alias{use_job_template}
\title{Use a job template to submit a job to the cluster}
\usage{
use_job_template(
  replace_tibble,
  template,
  file_dir = tempdir(),
  temp_ext = ".sh",
  temp_prefix = "job_",
  warning_label = "",
  submit = TRUE,
  job_scheduler = "slurm"
)
}
\arguments{
\item{replace_tibble}{A tibble with two columns, find and replace.
The find column should contain the placeholder text to be replaced and the
replace column should contain the text to replace it with.}

\item{template}{The name of the template file to use. This should be a file
in the rrrSingleCellUtils/inst folder. The job_scheduler argument will be
prefixed to the template name to find the file.}

\item{file_dir}{The directory to write the temporary job file to.}

\item{temp_ext}{The extension to use for the temporary job file.}

\item{temp_prefix}{The prefix to use for the temporary job file.}

\item{warning_label}{A string to use in the warning message if the job
submission fails.}

\item{submit}{Whether to submit the sbatch jobs to the cluster or not.}

\item{job_scheduler}{The job scheduler to use. One of "slurm", "sge" or
"bash".}
}
\value{
0 if the job submission was successful, otherwise an error is
 thrown
}
\description{
Use a job template to submit a job to the cluster
}
