% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda.R
\name{confirm_conda_env}
\alias{confirm_conda_env}
\title{confirm_conda_env}
\usage{
confirm_conda_env()
}
\description{
confirm_conda_env
}
\details{
Checks if the conda environment "scanBit_xkcd_1337" exists and
creates it if not.
}
