% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_call.R
\name{get_snp_tree}
\alias{get_snp_tree}
\title{Generate a tree using SNPs derived from single cell clusters}
\usage{
get_snp_tree(
  cellid_bam_table,
  temp_dir = tempdir(),
  output_dir,
  output_base_name = "hier_tree",
  log_base = "jobOut",
  job_base = "sbatch_",
  account = "gdrobertslab",
  ploidy,
  ref_fasta,
  min_depth = 5,
  min_snvs_per_cluster = 500,
  max_prop_missing_at_site = 0.9,
  n_bootstraps = 10000,
  bootstrap_cutoff = 0.99,
  verbose = TRUE,
  submit = TRUE,
  cleanup = TRUE,
  other_job_header_options = "",
  other_batch_options = "",
  job_scheduler = "slurm"
)
}
\arguments{
\item{cellid_bam_table}{A tibble with three columns: cell_barcode,
cell_group and bam_file. The cell_barcode column should contain the cell
barcode, the cell_group column should contain the cluster label and the
bam_file column should contain the path to the bam file for that cell.}

\item{temp_dir}{The directory to write temporary files to.}

\item{output_dir}{The directory to write output distance files to.}

\item{output_base_name}{The prefix to use with the output files.}

\item{log_base}{The directory to write job output files to.}

\item{job_base}{The prefix to use with the sbatch job file.}

\item{account}{The hpc account to use.}

\item{ploidy}{Either a file path to a ploidy file, or a string indicating
the ploidy. #### PUT EXAMPLES HERE #### For ploidy file format see bcftools
call --ploidy-file option description here:
"https://samtools.github.io/bcftools/bcftools.html"}

\item{ref_fasta}{The path to the reference fasta file.}

\item{min_depth}{The minimum depth to use when calling SNPs. Can be a vector
of numbers.}

\item{min_snvs_per_cluster}{The minimum number of sites that must be covered
by a cell_group to be included in the tree.}

\item{max_prop_missing_at_site}{The maximum proportion of missing data
allowed at a site.}

\item{n_bootstraps}{The number of bootstraps to use when evaluating grouping
of samples in the hierarchical clustering tree.}

\item{bootstrap_cutoff}{The bootstrap cutoff to use when collapsing the tree.
This is the proportion of bootstraps that must support a grouping for it to
be considered a valid grouping.}

\item{verbose}{Whether to print out verbose output or not.}

\item{submit}{Whether to submit the sbatch jobs to the cluster or not.}

\item{cleanup}{Whether to clean up the temporary files after execution.}

\item{other_job_header_options}{Other options to pass to the batch job
header. This should be a list of strings. Do not put "#SBATCH" or "#$" in
front of the strings.}

\item{other_batch_options}{Other options to pass to the batch job. These
optional strings will not be part of the header and will be placed verbatim
after the header. This is useful for passing module load commands, adding
to the path or other system-specific commands. Should be a list of strings.}

\item{job_scheduler}{The job scheduler to use. One of "slurm", "sge" or
"bash".}
}
\value{
TRUE if the function succeeded. Output is written to output_dir.
}
\description{
Generate a tree using SNPs derived from single cell clusters
}
\details{
for ploidy, GRCh37 is hg19, GRCh38 is hg38, X, Y, 1, mm10_hg19 is
  our mixed species reference with species prefixes on chromosomes, mm10 is
  mm10
}
\examples{
\dontrun{
placeholder for now
}
}
