% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_call.R
\name{group_clusters_by_dist}
\alias{group_clusters_by_dist}
\title{Use Phylogenetic Tree from Merged BCF File to Find Similar Clusters}
\usage{
group_clusters_by_dist(
  merged_bcf_file,
  output_file = paste0(merged_bcf_file, "_groups.txt"),
  min_snvs_per_cluster = 500,
  max_prop_missing_at_site = 0.9,
  n_bootstraps = 1000,
  bootstrap_cutoff = 0.99,
  dist_method,
  linkage_method,
  tree_figure_file,
  n_comps_file,
  verbose = TRUE,
  job_base = "job_dist",
  account = "gdrobertslab",
  log_base = "jobOut",
  temp_dir = tempdir(),
  submit = TRUE,
  other_job_header_options = "",
  other_batch_options = "",
  sge_q,
  sge_parallel_environment,
  job_scheduler = "slurm"
)
}
\arguments{
\item{merged_bcf_file}{Character. Path to the merged BCF file.}

\item{output_file}{Character. Path to the output file for assigned groupings
per sample (cluster). Default is merged_bcf_file + "_groups.txt".}

\item{min_snvs_per_cluster}{Numeric. Minimum number of SNVs per cluster.
Default is 500.}

\item{max_prop_missing_at_site}{The maximum proportion of missing data
allowed at a site.}

\item{n_bootstraps}{The number of bootstraps to use when evaluating grouping
of samples in the hierarchical clustering tree.}

\item{bootstrap_cutoff}{The bootstrap cutoff to use when collapsing the tree.
This is the proportion of bootstraps that must support a grouping for it to
be considered a valid grouping.}

\item{dist_method}{The distance method to use when calculating the distance
matrix. Can be "binary" or "euclidian". The binary method calculates the
proportion of sites that are different between two samples, while the
euclidian method calculates the euclidian distance between the samples
based on the proportion of reference alleles at each site. The default is
"binary".}

\item{linkage_method}{The linkage method to use for hierarchical clustering.
Can be "single", "complete", "average", "weighted", "centroid", "median",
or "ward". The default is "ward".}

\item{tree_figure_file}{Character. Path to the output tree figure.}

\item{n_comps_file}{Character. Path to the output file for number of
shared comparable SNVs between each pair of clusters.}

\item{verbose}{Whether to print out verbose output or not.}

\item{job_base}{The prefix to use with the sbatch job file.}

\item{account}{The hpc account to use.}

\item{log_base}{The directory to write job output files to.}

\item{temp_dir}{The directory to write temporary files to. This should be
unique to each sample run to avoid file collisions.}

\item{submit}{Whether to submit the sbatch jobs to the cluster or not.}

\item{other_job_header_options}{Other options to pass to the batch job
header. This should be a list of strings. Do not put "#SBATCH" or "#$" in
front of the strings.}

\item{other_batch_options}{Other options to pass to the batch job. These
optional strings will not be part of the header and will be placed verbatim
after the header. This is useful for passing module load commands, adding
to the path or other system-specific commands. Should be a list of strings.}

\item{sge_q}{Character string specifying the cluster queue when using sge as
the job scheduler.}

\item{sge_parallel_environment}{Character string specifying the parallel
environment (part of the -pe argument in the header) when using sge as the
job scheduler}

\item{job_scheduler}{The job scheduler to use. One of "slurm", "sge" or
"bash". If specifying "sge", sge_q and sge_parallel should be specified if
defaults are not correct.}
}
\value{
The result of the SLURM job submission.
}
\description{
This function calculates how samples (from pseudobulked clusters) from a
merged BCF file cluster hierarchically by collapsing a bootstrapped
hierarchical clustered tree. This provides mathematical testing of
genotypic divergence between the samples (clusters).
}
\examples{
\dontrun{
  group_clusters_by_dist(
      merged_bcf_file = "path/to/merged.bcf",
      tree_figure_file = "path/to/tree_figure.png"
  )
}

}
