% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_call.R
\name{group_clusters_by_dist}
\alias{group_clusters_by_dist}
\title{Use Phylogenetic Tree from Merged BCF File to Find Similar Clusters}
\usage{
group_clusters_by_dist(
  merged_bcf_file,
  output_file = paste0(merged_bcf_file, "_dist.txt"),
  min_snvs_per_cluster = 500,
  max_prop_missing_at_site = 0.9,
  n_bootstraps = 1000,
  bootstrap_cutoff = 0.99,
  tree_figure_file,
  verbose = TRUE,
  sbatch_base = "sbatch_dist",
  account = "gdrobertslab",
  slurm_base = "slurmOut",
  temp_dir = tempdir(),
  submit = TRUE,
  other_sbatch_options = ""
)
}
\arguments{
\item{merged_bcf_file}{Character. Path to the merged BCF file.}

\item{output_file}{Character. Path to the output file for assigned groupings
per sample (cluster). Default is merged_bcf_file + "_dist.txt".}

\item{min_snvs_per_cluster}{Numeric. Minimum number of SNVs per cluster.
Default is 500.}

\item{max_prop_missing_at_site}{The maximum proportion of missing data
allowed at a site.}

\item{n_bootstraps}{The number of bootstraps to use when evaluating grouping
of samples in the hierarchical clustering tree.}

\item{bootstrap_cutoff}{The bootstrap cutoff to use when collapsing the tree.
This is the proportion of bootstraps that must support a grouping for it to
be considered a valid grouping.}

\item{tree_figure_file}{Character. Path to the output tree figure.}

\item{verbose}{Whether to print out verbose output or not.}

\item{sbatch_base}{The prefix to use with the sbatch job file.}

\item{account}{The hpc account to use.}

\item{slurm_base}{The directory to write slurm output files to.}

\item{temp_dir}{The directory to write temporary files to. This should be
unique to each sample run to avoid file collisions.}

\item{submit}{Whether to submit the sbatch jobs to the cluster or not.}

\item{other_sbatch_options}{Other options to pass to the sbatch command. This
should be a list of strings. Do not put "#SBATCH" in front of the
strings.}
}
\value{
The result of the SLURM job submission.
}
\description{
This function calculates how samples (from pseudobulked clusters) from a
merged BCF file cluster hierarchically by collapsing a bootstrapped
hierarchical clustered tree. This provides mathematical testing of
genotypic divergence between the samples (clusters).
}
\examples{
\dontrun{
  group_clusters_by_dist(
      merged_bcf_file = "path/to/merged.bcf",
      tree_figure_file = "path/to/tree_figure.png"
  )
}

}
