% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_call.R
\name{group_clusters_by_dist}
\alias{group_clusters_by_dist}
\title{Use Phylogenetic Tree from Merged BCF File to Find Similar Clusters}
\usage{
group_clusters_by_dist(
  merged_bcf_file,
  output_file = paste0(merged_bcf_file, "_dist.txt"),
  min_snvs_per_cluster = 500,
  max_prop_missing_at_site = 0.9,
  n_bootstraps = 1000,
  bootstrap_cutoff = 0.99,
  tree_figure_file,
  verbose = TRUE,
  sbatch_base = "sbatch_dist",
  account = "gdrobertslab",
  slurm_base = "slurmOut",
  temp_dir,
  submit = TRUE
)
}
\arguments{
\item{merged_bcf_file}{Character. Path to the merged BCF file.}

\item{output_file}{Character. Path to the output file for assigned groupings
per sample (cluster). Default is merged_bcf_file + "_dist.txt".}

\item{min_snvs_per_cluster}{Numeric. Minimum number of SNVs per cluster.
Default is 500.}

\item{max_prop_missing_at_site}{Numeric. Maximum proportion of missing data
allowed at a variant site. Default is 0.9.}

\item{n_bootstraps}{Numeric. Number of bootstrap iterations. Default is 1000.}

\item{bootstrap_cutoff}{Numeric. Bootstrap cutoff threshold to use for
collapsing groups of samples (clusters). Default is 0.99.}

\item{tree_figure_file}{Character. Path to the output tree figure.}

\item{verbose}{Logical. If TRUE, enables verbose output. Default is TRUE.}

\item{sbatch_base}{Character. Base name for SLURM batch scripts. Can put
these in /tmp/ if you want. Default is "./sbatch_dist".}

\item{account}{Character. SLURM account name. Default is "gdrobertslab".}

\item{slurm_base}{Character. Path to the SLURM output file.}

\item{temp_dir}{The directory to write temporary files to.}

\item{submit}{Logical. If TRUE, submits the SLURM job. Default is TRUE. Used
mostly for testing.}
}
\value{
The result of the SLURM job submission.
}
\description{
This function calculates how samples (from pseudobulked clusters) from a
merged BCF file cluster hierarchically by collapsing a bootstrapped
hierarchical clustered tree. This provides mathematical testing of
genotypic divergence between the samples (clusters).
}
\examples{
\dontrun{
  group_clusters_by_dist(
      merged_bcf_file = "path/to/merged.bcf",
      tree_figure_file = "path/to/tree_figure.png"
  )
}

}
